+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased][unreleased]
 
+### Added
+
+- Allow service support users to upload GIAS data for asynchronous ingestion
+
 ## [Release 43][release-43]
 
 ### Fixed